+++ packages/vue/src/index.ts
@@ -1,4 +1,5 @@
 export { useAsyncData } from "./composables/UseAsyncData.js";
+export { useCookies } from "./composables/useCookies.js";
 export { definePrefetchStore } from "./definitions/DefinePrefetchStore.js";
 export { defineAsyncApp } from "./definitions/DefineAsyncApp.js";
 export { createApp } from "./utils/CreateApp.js";