+++ packages/_example-ts-vue/CHANGELOG.md
@@ -1,5 +1,17 @@
 # example-ts-vue
 
+## 0.0.11
+
+### Patch Changes
+
+- Add cookies supply
+- Updated dependencies
+  - lite-ssr@0.4.7
+  - @lite-ssr/vue@0.4.7
+  - @lite-ssr/proxy@1.0.6
+  - @lite-ssr/cached-data@1.0.2
+  - @lite-ssr/vue-unhead@1.0.8
+
 ## 0.0.10
 
 ### Patch Changes