+++ packages/cached-data/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @lite-ssr/cached-data
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @lite-ssr/vue@0.4.7
+
 ## 1.0.1
 
 ### Patch Changes