+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lite-ssr/core
 
+## 0.4.4
+
+### Patch Changes
+
+- Add cookies supply
+
 ## 0.4.3
 
 ### Patch Changes