+++ packages/lite-ssr/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.4.7
+
+### Patch Changes
+
+- Add cookies supply
+- Updated dependencies
+  - @lite-ssr/core@0.4.4
+
 ## 0.4.6
 
 ### Patch Changes