+++ packages/vue-unhead/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @lite-ssr/vue
 
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @lite-ssr/vue@0.4.7
+
 ## 1.0.7
 
 ### Patch Changes